lab3/requirements.txt
numpy
hydra-core
gym
tqdm
wandb
prettytable==3.2.0
Pillow==9.0.0